requirements.txt
aiogram 
aiohttp 
apscheduler 
pyyaml 
python-dotenv